+++ src/policy/packages.cpp
@@ -104,3 +104,23 @@ bool IsChildWithParents(const Package &package) {
                            return input_txids.count(ptx->GetId()) > 0;
                        });
 }
+
+bool IsChildWithParentsTree(const Package &package) {
+    if (!IsChildWithParents(package)) {
+        return false;
+    }
+    std::unordered_set<TxId, SaltedTxIdHasher> parent_txids;
+    std::transform(package.cbegin(), package.cend() - 1,
+                   std::inserter(parent_txids, parent_txids.end()),
+                   [](const auto &ptx) { return ptx->GetId(); });
+    // Each parent must not have an input who is one of the other parents.
+    return std::all_of(
+        package.cbegin(), package.cend() - 1, [&](const auto &ptx) {
+            for (const auto &input : ptx->vin) {
+                if (parent_txids.count(input.prevout.GetTxId()) > 0) {
+                    return false;
+                }
+            }
+            return true;
+        });
+}