+++ src/rpc/mempool.cpp
@@ -782,18 +782,16 @@ static RPCHelpMan submitpackage() {
     return RPCHelpMan{
         "submitpackage",
         "Submit a package of raw transactions (serialized, hex-encoded) to "
-        "local node (-regtest only).\n"
+        "local node.\n"
+        "The package must consist of a child with its parents, and none of the "
+        "parents may depend on one another.\n"
         "The package will be validated according to consensus and mempool "
         "policy rules. If any transaction passes, it will be accepted to "
         "mempool.\n"
         "This RPC is experimental and the interface may be unstable. Refer to "
         "doc/policy/packages.md for documentation on package policies.\n"
-        "Warning: until package relay is in use, successful submission does "
-        "not mean the transaction will propagate to other nodes on the "
-        "network.\n"
-        "Currently, each transaction is broadcasted individually after "
-        "submission, which means they must meet other nodes' feerate "
-        "requirements alone.\n",
+        "Warning: successful submission does not mean the transactions will "
+        "propagate throughout the network.\n",
         {
             {
                 "package",
@@ -856,10 +854,6 @@ static RPCHelpMan submitpackage() {
                     HelpExampleCli("submitpackage", "[rawtx1, rawtx2]")},
         [&](const RPCHelpMan &self, const Config &config,
             const JSONRPCRequest &request) -> UniValue {
-            if (!Params().IsMockableChain()) {
-                throw std::runtime_error("submitpackage is for regression "
-                                         "testing (-regtest mode) only");
-            }
             const UniValue raw_transactions = request.params[0].get_array();
             if (raw_transactions.size() < 1 ||
                 raw_transactions.size() > MAX_PACKAGE_COUNT) {
@@ -880,6 +874,12 @@ static RPCHelpMan submitpackage() {
                             " Make sure the tx has at least one input.");
                 }
                 txns.emplace_back(MakeTransactionRef(std::move(mtx)));
+            }
+            if (!IsChildWithParentsTree(txns)) {
+                throw JSONRPCTransactionError(
+                    TransactionError::INVALID_PACKAGE,
+                    "package topology disallowed. not child-with-parents or "
+                    "parents depend on each other.");
             }
 
             NodeContext &node = EnsureAnyNodeContext(request.context);
@@ -1011,7 +1011,7 @@ void RegisterMempoolRPCCommands(CRPCTable &t) {
         {"blockchain", getmempoolinfo},
         {"blockchain", getrawmempool},
         {"blockchain", savemempool},
-        {"hidden", submitpackage},
+        {"rawtransactions", submitpackage},
     };
     for (const auto &c : commands) {
         t.appendCommand(c.name, &c);